requirements.txt
langchain
python-dotenv
langchain-community
ipykernel
beautifulsoup4
faiss-cpu
streamlit


